+++ addons/clyde/examples/examples_with_helpers/fixed_bubble/example_fixed_with_helpers.gd
@@ -1,5 +1,7 @@
 extends MarginContainer
 
+# NOTE This example requires the helpers option to be enabled in Project Settings.
+
 @onready var _dialogue_start_button := $VBoxContainer/Button
 @onready var _dialogue_selector := $VBoxContainer/OptionButton
 
@@ -13,7 +15,12 @@ var _dialogue_files = [
 
 var _current_dialogue = 0
 
-# NOTE This example requires the helpers option to be enabled in Project Settings.
+# NOTE: This Dialogue overwrite is only required for these examples to prevent
+# startup errors when the Dialogue helper is not enabled.
+# Feel free to remove this line if you have the helpers enabled
+@onready
+var Dialogue = get_tree().root.get_node("Dialogue") if get_tree().root.has_node("Dialogue") else {}
+
 func _ready():
 	_setup_dialogue_events()
 	_load_buttons()